fix time validator for positive numbers that were less than 1

+++ src/duckling/game/path/path.component.ts
@@ -88,7 +88,7 @@ export class PathComponent {
 
     get timeValidator() : Validator {
         return (value : string) => {
-            return Number.isInteger(parseInt(value)) && parseInt(value) > 0;
+            return Number.isInteger(parseInt(value)) && parseFloat(value) > 0;
         }
     }
 }
